major edits
