+++ frontend/src/components/SyncEditor/SyncEditor.tsx
@@ -14,9 +14,9 @@ import {
     CloseIcon,
 } from '@patternfly/react-icons/dist/js/icons'
 import { ClipboardCopyButton } from '@patternfly/react-core'
-import Ajv from 'ajv'
 import { debounce, noop, isEqual, cloneDeep } from 'lodash'
-import { processForm, processUser, formatErrors, ProcessedType } from './process'
+import { processForm, processUser, ProcessedType } from './process'
+import { compileAjvSchemas, formatErrors } from './validation'
 import { getFormChanges, getUserChanges, formatChanges } from './changes'
 import { decorate, getResourceEditorDecorations } from './decorate'
 import { SyncDiffType } from './SyncDiff'
@@ -86,33 +86,7 @@ export function SyncEditor(props: SyncEditorProps): JSX.Element {
     // compile schema(s) just once
     const validationRef = useRef<unknown>()
     if (schema && !validationRef.current) {
-        try {
-            const ajv = new Ajv({ allErrors: true, verbose: true })
-            ajv.addKeyword({
-                keyword: 'validateDNSName',
-                schemaType: 'boolean',
-                validate: (_schema: null, data: any) => {
-                    return (
-                        !data ||
-                        (/^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$/.test(data) &&
-                            data.length <= 63)
-                    )
-                },
-            })
-            if (!Array.isArray(schema)) {
-                validationRef.current = [{ validator: ajv.compile(schema) }]
-            } else {
-                const schemas: any = []
-                schema.forEach(({ type, required, schema }) => {
-                    schemas.push({
-                        type,
-                        required,
-                        validator: ajv.compile(schema),
-                    })
-                })
-                validationRef.current = schemas
-            }
-        } catch (e) {}
+        validationRef.current = compileAjvSchemas(schema)
     }
 
     function onEditorDidMount(editor: any, monaco: any) {
@@ -236,11 +210,20 @@ export function SyncEditor(props: SyncEditorProps): JSX.Element {
                     const editor = editorRef.current
                     const model = editor.getModel()
                     const pos = editor.getPosition()
+                    const lines = model.getLineCount()
                     const thisLine = model.getLineContent(pos.lineNumber)
-                    const nextLine = model.getLineContent(pos.lineNumber + 1)
-                    const times = Math.max(thisLine.search(/\S/), nextLine.search(/\S/))
+                    let times
+                    const isLastLine = lines <= pos.lineNumber + 1
+                    if (isLastLine) {
+                        times = thisLine.search(/\S/)
+                    } else {
+                        const nextLine = model.getLineContent(pos.lineNumber + 1)
+                        times = Math.max(thisLine.search(/\S/), nextLine.search(/\S/))
+                    }
                     const count = `${newKeyCount}`.padStart(4, '0')
-                    const newLine = `${' '.repeat(times)}key${count}:  \n`
+                    const newLine = `${isLastLine ? '\n' : ''}${' '.repeat(times)}key${count}:  ${
+                        !isLastLine ? '\n' : ''
+                    }`
                     let range = new monacoRef.current.Range(pos.lineNumber + 1, 0, pos.lineNumber + 1, 0)
                     editor.executeEdits('new-key', [{ identifier: 'new-key', range, text: newLine }])
                     range = new monacoRef.current.Range(pos.lineNumber + 1, times + 1, pos.lineNumber + 1, times + 8)
